:recycle: Refactor structura
