+++ vinilos_g24/app/src/main/java/com/misw/vinilos_g24/repositories/AlbumRepository.kt
@@ -1,0 +1,6 @@
+package com.misw.vinilos_g24.repositories
+
+import android.app.Application
+class AlbumRepository (val application: Application){
+    fun refreshData() {}
+}
