--- vinilos_g24/app/src/main/java/com/misw/vinilos_g24/ui/albumes/AlbumesViewModel.kt
@@ -1,13 +1,0 @@
-package com.misw.vinilos_g24.ui.albumes
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class AlbumesViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "Álbumes"
-    }
-    val text: LiveData<String> = _text
-}
