+++ vinilos_g24/app/src/main/java/com/misw/vinilos_g24/ui/home/HomeFragment.kt
@@ -8,6 +8,7 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import com.misw.vinilos_g24.databinding.FragmentHomeBinding
+import com.misw.vinilos_g24.viewmodels.HomeViewModel
 
 class HomeFragment : Fragment() {
