+++ vinilos_g24/app/src/main/java/com/misw/vinilos_g24/viewmodels/ArtistasViewModel.kt
@@ -1,13 +1,13 @@
-package com.misw.vinilos_g24.ui.home
+package com.misw.vinilos_g24.viewmodels
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class HomeViewModel : ViewModel() {
+class ArtistasViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "Inicio"
+        value = "Artistas"
     }
     val text: LiveData<String> = _text
 }
