+++ vinilos_g24/app/src/main/java/com/misw/vinilos_g24/viewmodels/HomeViewModel.kt
@@ -1,13 +1,13 @@
-package com.misw.vinilos_g24.ui.artistas
+package com.misw.vinilos_g24.viewmodels
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class ArtistasViewModel : ViewModel() {
+class HomeViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "Artistas"
+        value = "Inicio"
     }
     val text: LiveData<String> = _text
 }
